+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New arg **--source-sub-title**, to define the **Source** image subtitles (defaults to 'Source').
+- New arg **--fpng-compression**, this allows control over fpng's compression level (defaults to 1 - slow compression).
 
 ### Changed